*修改传参类型

+++ golis.go
@@ -192,13 +192,13 @@ func Int64ToBytes(n int64) []byte {
 }
 
 //字节转换成整形64
-func BytesToInt64(b []byte) int {
+func BytesToInt64(b []byte) int64 {
 	bytesBuffer := bytes.NewBuffer(b)
 
 	var x int64
 	binary.Read(bytesBuffer, binary.BigEndian, &x)
 
-	return int(x)
+	return int64(x)
 }
 
 //整形64转换成字节
@@ -211,13 +211,13 @@ func Int8ToBytes(n int8) []byte {
 }
 
 //字节转换成整形8
-func BytesToInt8(b []byte) int {
+func BytesToInt8(b []byte) int8 {
 	bytesBuffer := bytes.NewBuffer(b)
 
 	var x int8
 	binary.Read(bytesBuffer, binary.BigEndian, &x)
 
-	return int(x)
+	return int8(x)
 }
 
 //简单日志输出
